requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-extensions==3.2.3
django-filter==23.3
drf-spectacular==0.26.4
drf-spectacular-sidecar==2023.10.1
whitenoise==6.6.0
pytest==7.4.3
pytest-django==4.7.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
pandas==2.1.1
openpyxl==3.1.2 